Prevent server status processing on initial form submission
Previously, when submitting the form, the status method ran again and tested the server as specified in the path that the form was initially built on. This change checks to see if the form was submitted, and if so, we just render the form and don't do any server testing, because that is done once the form redirects, and uses the route parameter to determine which server to test.